pyproject.toml
[tool.poetry]
name = "sudimir-backend"
version = "0.1.0"
description = "Backend part for MEPFI APP"
authors = ["Dermofet"]


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.0"
SQLAlchemy = "^2.0.17"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
asyncpg = "^0.27.0"
pydantic = "^2.3.0"
loguru = "^0.7.0"
alembic = "^1.11.0"
jwt = "^1.3.1"
phonenumbers = "^8.13.15"
passlib = "^1.7.4"
python-jose = "^3.3.0"
pydantic-settings = "^2.0.3"


[tool.poetry.dev-dependencies]
mypy = "^0.942"
black = {extras = ["d"], version = "^22.8.0"}
isort = "^5.10.1"
pytest = "^7.1.1"
sqlalchemy-stubs = "^0.4"
pytest-cov = "^3.0.0"
pytest-alembic = "^0.8.4"
pytest-asyncio = "^0.19.0"
pytest-trio = "^0.7.0"


[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"